+++ src/string.js
@@ -27,3 +27,10 @@ export let toSentence = _.flow(
   intersperse(differentLast(() => ', ', () => ' and ')),
   _.join('')
 )
+
+export let toSentenceWith = _.curry((x, array) =>
+  _.flow(
+    intersperse(differentLast(() => ', ', () => x)),
+    _.join('')
+  )(array)
+)